Allow non-TTY stdin watch mode (#448)
The presence of stdin doesn't necessarily mean there's an allocated
tty. This breaks watch mode in non-TTY stdin contexts (e.g. docker,
foreman, etc). A simple process.stdin.isTTY check would theoretically
be enough but unfortunately, subprocesses don't have the same API,
which are used extensively to test via calls to `spawn`.

A simple solution is to inject an env var dependency where we tell
the process that it's indeed a TTY-allocated process and so, watch mode
with exit handling is good to go.

Co-authored-by: Thiago Brandão <194487+thiagobrandam@users.noreply.github.com>